Break out not.hpp
