+++ test/compiling_tests_14.cpp
@@ -2875,7 +2875,7 @@ Tried obj.overload\(trompeloeil::re<std::string const&>\("end\$", std::regex_con
 
 #endif /* TROMPELOEIL_TEST_REGEX_BOL_EOL_FAILURES */
 
-// tests of parameter matching using neg_matcher
+// tests of parameter matching using not_matcher
 
 TEST_CASE_METHOD(
   Fixture,
